feat(data): serve static fixtures in prod demo (#15)
